depth_ftn_type
- Added a quicker way to compute depths of newly generated points. Now
the chull points' dot product with each unit vector and the sorting are
done once. So while testing if a point is inside it or not, the dot
product and sorting are not done every time there arent enough points
inside the chull.
- Added min_pts_in_chull variable. With this variable, if points
generated inside a chull are lesser than it per iteration then after
max_chull_tries, the optimization terminates. Previously, a check was
made that at least one point is inside the chull. This can be very
inefficient sometimes.

- The optimization terminated without any trace a bunch of times. Upon
using the boundscheck flag, it disappeared. Translations of the C
function into cython might help to find it. So far, it seems to happen
after getting out of the optimization loop. Maybe in the last call to
get_new_chull_vecs?