Fixed ROPE (kinda)
- Instead of checking (in get_new_chull_vecs) if the best points are
boundary or not, all points are taken. This keeps the best points and
the objective function keeps on getting better instead of wandering
around. This happend because the best points were inside the hull and
not taken.

- One the first attempt in gen_vecs_in_chull, all chull_vecs (or
acc_vecs) are kept and rest of the n_temp_rope_prm_vecs are generated.
For the next attempts all points are randomly generated and tested for
containment.

- The shrinkage constant for the depth function is increased from 1e-10
to 1e-6.

- Sometimes, a point that is supposed to be inside or on the boundary,
gets a depth of zero. This happens randomly in the beginning iterations
(around 3). More so on Threadripper. This certainly has to do something
with the depth function and the shrinkage constant.

- Sometimes the optimization goes into in an infinite loop apparently.
This has happend only twice till now (only on Threadripper). Cannot pin
it. The only while loops that can do this are in the sorting functions
(that I am not very good at understanding).